third_party/WebKit/LayoutTests/fast/borders/border-image-width-negative.html

<!doctype html>
<html>
  <head>
    <script src="../../resources/js-test.js"></script>
    <style>
    .test {
        border-image-width: 2 3 4 5;
        border-image-width: -1;
        border-image-width: -1 2;
        border-image-width: 1 2 -0.1;
        border-image-width: 1 2 3 -0.1;

        border-image-outset: 6 7 8 9;
        border-image-outset: -1;
        border-image-outset: -1 2;
        border-image-outset: 1 2 -0.1;
        border-image-outset: 1 2 3 -0.1;

        border-style: solid;
        visibility: hidden;
    }
    </style>
  </head>
  <body>
    <div class="test"></div>
    <script>
    description("Test that negative border image width values are ignored");
    var elm = document.querySelector('.test');
    shouldBeEqualToString("getComputedStyle(elm).borderImageWidth", "2 3 4 5");
    shouldBeEqualToString("getComputedStyle(elm).borderImageOutset", "6 7 8 9");
    </script>
  </body>
</html>
