third_party/WebKit/LayoutTests/fast/borders/border-image-width-negative-expected.txt

Test that negative border image width values are ignored

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(elm).borderImageWidth is "2 3 4 5"
PASS getComputedStyle(elm).borderImageOutset is "6 7 8 9"
PASS successfullyParsed is true

TEST COMPLETE
